feat: implement 'dispatch namespaces'
This feature is for 'Workers for Platforms', and enables usage of a few
'namespaces' APIs to get the bare minimum example running.

New Commands:
  wrangler dispatch-namespace create <name>
  wrangler dispatch-namespace delete <name>

New Bindings:
  [[dispatch_namespaces]]
  name = "dispatcher"
  namespace = "my-namespace"

New Config:
  dispatch_namespace = "my-namespace"

Example usage would be to create a namespace with `dispatch-namespace
create my-namespace`, upload a worker *to* that namespace by including
`dispatch_namespace = "my-namespace"` in a worker's toml, and then
upload a worker with a 'namespace' binding that can be used to dispatch
to the worker we just uploaded (by including a `[[dispatch_namespaces]]`
binding in its `wrangler.toml`).

After doing those three things, you should be able to route to the last
worker and call out to the first worker using the new binding.